* master:
  update and fix testvectors (#146)
  validator service tests (#145)
  safrole service tests (#143)
  fix log filtering (#142)
  improve logging (#141)
  refactor and more tests (#140)
  Block author tickets (#139)
  block author tests (#138)
  Fix accessor (#137)
  tests for DispatchQueueScheduler (#136)
  update readme
  use external libs (#133)
  use saturating number for gas and balance (#135)
  pvm accumulate functions and inovcation (#132)

+++ Blockchain/Sources/Blockchain/Validator/DevKeyStore.swift
@@ -4,11 +4,11 @@ import Utils
 public final class DevKeyStore: KeyStore {
     private let keystore: InMemoryKeyStore
 
-    public init(devKeysCount: UInt32 = 12) async throws {
+    public init(devKeysCount: Int = 12) async throws {
         keystore = InMemoryKeyStore()
 
         for i in 0 ..< devKeysCount {
-            _ = try await addDevKeys(seed: i)
+            _ = try await addDevKeys(seed: UInt32(i))
         }
     }
